Add files via upload
Beta Calibration: The code optimizes the beta parameter to better fit the actual data of infected individuals over time. The loss function compares the predicted infected count with the actual observed data, and the scipy.optimize.minimize_scalar() function is used to find the best value of beta that minimizes this error.